Implement atomic incrementing index for qnService
We need an unique, automatically incrementing id for each question
created in the database. To do so in mongodb, we create a counter
document in mongoDB, and increment the count in that document each time
we create a new question document.

This ensure atomicity since there can only be one document fetching and
updating the counter document at any one time. This also ensures
uniqueness of the id assigned to each question.